# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Debt:
    properties:
      amount:
        description: cents
        type: integer
      from_user_id:
        type: integer
      to_user_id:
        type: integer
    type: object
...
  domain.Settlement:
    properties:
      amount:
        description: cents
        type: integer
      created_at:
        type: string
      from_user:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Relationships
      from_user_id:
        type: integer
      group_id:
        type: integer
      id:
        type: integer
      to_user:
        $ref: '#/definitions/domain.User'
      to_user_id:
        type: integer
    type: object
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      push_enabled:
        description: |-
          PushEnabled is the user's own preference, checked before sending any
          push notification — separate from whether they have any subscribed
          devices (PushSubscription rows).
        type: boolean
      updated_at:
        type: string
    type: object
...
  handler.SettleDebtRequest:
    properties:
      amount:
        example: 5000
        type: integer
      from_user_id:
        example: 2
        type: integer
      to_user_id:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for SplitEasy app (Splitwise style expense sharing).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SplitEasy API
  version: "1.0"
paths:
...
  /groups/{id}/balances:
    get:
      description: Calculates the debts within a group to settle accounts.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Debt'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Get balances/debts of a group
      tags:
      - groups
...
  /groups/{id}/settlements:
    get:
      description: Returns every recorded payment in the group for the unified history
        view.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Settlement'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: List a group's settlements
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Records a payment between two group members, reducing their outstanding
        balance.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Settlement Info
        in: body
        name: settlement
        required: true
        schema:
          $ref: '#/definitions/handler.SettleDebtRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Settlement'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Settle a debt
      tags:
      - groups
...
  /settlements/{id}:
    delete:
      description: Soft-deletes a recorded payment. Only the settlement's from_user_id
        or to_user_id may delete it — not just any group member.
      parameters:
      - description: Settlement ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Delete a settlement
      tags:
      - groups
    get:
      description: Retrieves a single settlement by id. Any member of the settlement's
        group may view it — unlike delete, this isn't limited to the two parties.
      parameters:
      - description: Settlement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Settlement'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Get a settlement
      tags:
      - groups
...
securityDefinitions:
  JWT:
    description: Type "Bearer <your-jwt-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"